+++ CHANGELOG.md
@@ -7,6 +7,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.6.7] 2022-06-03
+
+### Added
+
+- **IMaterial** now is cloneable entity.
+
+### Changed 
+
+### Fixed
+
 ## [0.6.6] 2022-06-02
 
 ### Added